added tests for scroll event filtering, refactored widget_input and moved tests to separate file
